refactor: entity 접미어 적용
