+++ backend/src/main/java/com/zzang/chongdae/offering/repository/entity/OfferingEntity.java
@@ -1,13 +1,16 @@
-package com.zzang.chongdae.offering.repository;
+package com.zzang.chongdae.offering.repository.entity;
 
 import com.zzang.chongdae.common.repository.BaseTimeEntity;
 import com.zzang.chongdae.member.repository.Member;
+import com.zzang.chongdae.offering.domain.OfferingPrice;
+import com.zzang.chongdae.offering.domain.OfferingStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -21,8 +24,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EqualsAndHashCode(of = "id", callSuper = false)
+@Table(name = "offering")
 @Entity
-public class Offering extends BaseTimeEntity {
+public class OfferingEntity extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -60,4 +64,12 @@ public class Offering extends BaseTimeEntity {
 
     @NotNull
     private BigDecimal totalPrice;
+
+    public OfferingPrice toOfferingPrice() {
+        return new OfferingPrice(totalCount, totalPrice);
+    }
+
+    public OfferingStatus toOfferingStatus(int currentCount) {
+        return new OfferingStatus(deadline, totalCount, isManualConfirmed, currentCount);
+    }
 }
