+++ backend/src/main/java/com/zzang/chongdae/offeringmember/repository/entity/OfferingMemberEntity.java
@@ -1,13 +1,17 @@
-package com.zzang.chongdae.offeringmember.repository;
+package com.zzang.chongdae.offeringmember.repository.entity;
 
 import com.zzang.chongdae.common.repository.BaseTimeEntity;
 import com.zzang.chongdae.member.repository.Member;
-import com.zzang.chongdae.offering.repository.Offering;
+import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
+import com.zzang.chongdae.offeringmember.domain.OfferingMemberRole;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -18,8 +22,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EqualsAndHashCode(of = "id", callSuper = false)
+@Table(name = "offering_member")
 @Entity
-public class OfferingMember extends BaseTimeEntity {
+public class OfferingMemberEntity extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,7 +34,8 @@ public class OfferingMember extends BaseTimeEntity {
     private Member member;
 
     @ManyToOne
-    private Offering offering;
+    private OfferingEntity offering;
 
+    @Enumerated(EnumType.STRING)
     private OfferingMemberRole role;
 }
